XWIKI-5880: When there are deleted attachments with id numbers greater than Integer.MAX_VALUE, XWiki#getDeletedAttachment(String) will fail or load incorrect attachments. Backport of r33934
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/platform/core/branches/xwiki-core-2.7@33935 f329d543-caf0-0310-9063-dda96c69346f

+++ xwiki-core/src/main/java/com/xpn/xwiki/XWiki.java
@@ -1697,7 +1697,7 @@ public class XWiki implements XWikiDocChangeNotificationInterface, EventListener
     public DeletedAttachment getDeletedAttachment(String id, XWikiContext context) throws XWikiException
     {
         if (hasAttachmentRecycleBin(context)) {
-            return getAttachmentRecycleBinStore().getDeletedAttachment(NumberUtils.toInt(id), context, true);
+            return getAttachmentRecycleBinStore().getDeletedAttachment(NumberUtils.toLong(id), context, true);
         }
         return null;
     }
